combine two function segment.
