+++ shardingsphere-test/shardingsphere-parser-test/src/main/java/org/apache/shardingsphere/test/sql/parser/parameterized/asserts/statement/dal/impl/ShowProcedureCodeStatementAssert.java
@@ -21,10 +21,9 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal.MySQLShowProcedureCodeStatement;
 import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.SQLCaseAssertContext;
+import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.segment.expression.ExpressionAssert;
 import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.statement.dal.ShowProcedureCodeStatementTestCase;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
 
 /**
  * MySQL show procedure code statement assert.
@@ -40,11 +39,6 @@ public final class ShowProcedureCodeStatementAssert {
      * @param expected      expected show procedure code statement test case
      */
     public static void assertIs(final SQLCaseAssertContext assertContext, final MySQLShowProcedureCodeStatement actual, final ShowProcedureCodeStatementTestCase expected) {
-        assertThat(assertContext.getText("Actual function name assertion error: "), actual.getFunction().getFunctionName(), is(expected.getFunction().getName()));
-        assertThat(assertContext.getText("Actual start index assertion error: "), actual.getFunction().getStartIndex(), is(expected.getFunction().getStartIndex()));
-        assertThat(assertContext.getText("Actual stop index assertion error: "), actual.getFunction().getStopIndex(), is(expected.getFunction().getStopIndex()));
-        if (null != expected.getFunction().getOwner()) {
-            assertThat(assertContext.getText("Actual owner assertion error: "), actual.getFunction().getOwner(), is(expected.getFunction().getOwner().getName()));
-        }
+        ExpressionAssert.assertFunction(assertContext, actual.getFunction(), expected.getFunction());
     }
 }
