+++ shardingsphere-test/shardingsphere-parser-test/src/main/resources/case/dal/show.xml
@@ -88,11 +88,11 @@
     <show-procedure-status sql-case-id="show_procedure_status" />
     <show-replicas sql-case-id="show_replicas" />
     <show-procedure-code sql-case-id="show_procedure_code_func">
-        <function name="TEST_FUNC" start-index="20" stop-index="28" />
+        <function function-name="TEST_FUNC" start-index="20" stop-index="28" text="TEST_FUNC" />
     </show-procedure-code>
     <show-procedure-code sql-case-id="show_procedure_code_owner_func">
-        <function name="TEST_FUNC" start-index="20" stop-index="38">
-            <owner name="TEST_USER" />
+        <function function-name="TEST_FUNC" start-index="20" stop-index="38" text="TEST_USER.TEST_FUNC">
+            <owner name="TEST_USER" start-index="20" stop-index="28"/>
         </function>
     </show-procedure-code>
 </sql-parser-test-cases>
